:hammer: Rewrite tests
