--- .jshintrc
@@ -1,14 +1,0 @@
-{
-  "strict": true,
-  "undef": true,
-  "unused": true,
-  "node": true,
-  "globals"   : {
-    "describe"   : false,
-    "it"         : false,
-    "before"     : false,
-    "beforeEach" : false,
-    "after"      : false,
-    "afterEach"  : false
-  }
-}
