+++ package.json
@@ -33,7 +33,7 @@
     }
   },
   "engines": {
-    "node": ">=4.3.0"
+    "node": ">=4.4.0"
   },
   "license": "MIT",
   "author": {
