+++ test/function-test.js
@@ -1,272 +1,22 @@
 'use strict';
-var config = require('./config'),
-  AV = require('..'),
-  should = require('should'),
-  fs = require('fs'),
-  request = require('supertest'),
-  assert = require('assert'),
-  _ = require('underscore');
 
-var appId = config.appId;
-var appKey = config.appKey;
-var masterKey = config.masterKey;
+const request = require('supertest');
+const should = require('should');
+const assert = require('assert');
 
-AV.init(config);
+const AV = require('..');
+const utils = require('./utils')
+const appInfo = require('./utils/app-info');
 
-var app;
+require('./utils/functions');
+require('./utils/hooks');
 
-if (process.env.FRAMEWORK == 'koa') {
-  var koa = require('koa')();
-  koa.use(AV.koa());
-  app = koa.listen();
-} else {
-  app = AV.express();
-}
+const appId = appInfo.appId;
+const appKey = appInfo.appKey;
+const masterKey = appInfo.masterKey;
+const sessionTokenAdmin = appInfo.sessionTokenAdmin;
 
-var TestObject = AV.Object.extend('TestObject');
-var ComplexObject = AV.Object.extend('ComplexObject');
-
-AV.Cloud.define('foo', function(request, response) {
-  assert.ok(request.meta.remoteAddress);
-  response.success("bar");
-});
-
-AV.Cloud.define('hello', function(request, response) {
-  response.success({action: "hello", name: request.params.name});
-});
-
-AV.Cloud.define('choice', function(req, res) {
-  if (req.params.choice) {
-    res.success('OK~');
-  } else {
-    res.error('OMG...');
-  }
-});
-
-AV.Cloud.define('complexObject', function(request, response) {
-  var query = new AV.Query(ComplexObject);
-  query.include('fileColumn');
-  query.ascending('createdAt');
-  query.find().then(
-    function(results) {
-      response.success({
-        foo: 'bar',
-        i: 123,
-        obj: {
-          a: 'b',
-          as: [1, 2, 3],
-        },
-        t: new Date('2015-05-14T09:21:18.273Z'),
-        avObject: results[0],
-        avObjects: results,
-      });
-    }
-  );
-});
-
-AV.Cloud.define('bareAVObject', function(request, response) {
-  var query = new AV.Query(ComplexObject);
-  query.include('fileColumn');
-  query.ascending('createdAt');
-  query.find().then(
-    function(results) {
-      response.success(results[0]);
-    }
-  );
-});
-
-AV.Cloud.define('AVObjects', function(request, response) {
-  var query = new AV.Query(ComplexObject);
-  query.include('fileColumn');
-  query.ascending('createdAt');
-  query.find().then(
-    function(results) {
-      response.success(results);
-    }
-  );
-});
-
-AV.Cloud.define('testAVObjectParams', function(request, response) {
-  request.params.avObject.should.be.instanceof(AV.Object);
-  request.params.avObject.get('name').should.be.equal('avObject');
-  request.params.avObject.get('pointerColumn').should.be.instanceof(AV.User);
-
-  request.params.avFile.should.be.instanceof(AV.File);
-
-  request.params.avObjects.forEach(function(object) {
-    object.should.be.instanceof(AV.Object);
-    object.get('name').should.be.equal('avObjects');
-  });
-
-  response.success();
-});
-
-AV.Cloud.define('testBareAVObjectParams', function(request, response) {
-  request.params.should.be.instanceof(AV.Object);
-  request.params.get('name').should.be.equal('avObject');
-  request.params.get('avFile').should.be.instanceof(AV.File);
-  request.params.get('avFile').name().should.be.equal('hello.txt');
-  response.success();
-});
-
-AV.Cloud.define('testAVObjectsArrayParams', function(request, response) {
-  request.params.forEach(function(object) {
-    object.get('name').should.be.equal('avObject');
-    object.get('avFile').should.be.instanceof(AV.File);
-    object.get('avFile').name().should.be.equal('hello.txt');
-  });
-  response.success();
-});
-
-AV.Cloud.define('testUser', function(request, response) {
-  assert.equal(request.user.className, '_User');
-  assert.equal(request.user.id, '54fd6a03e4b06c41e00b1f40');
-  assert.equal(request.user.get('username'), 'admin');
-  response.success('ok');
-});
-
-AV.Cloud.define('dontFetchUser', {fetchUser: false}, function(req, res) {
-  should.not.exist(res.user);
-  should.not.exist(res.currentUser);
-  req.sessionToken.should.be.equal(sessionToken_admin);
-  res.success();
-});
-
-AV.Cloud.define('testRun', function(request, response) {
-  if (request.params.shouldRemote && process.env.NODE_ENV != 'production') {
-    return response.error('Should be run on remote');
-  }
-
-  AV.Cloud.run('hello', {name: '李四'}).then(
-    function(data) {
-      assert.deepEqual(data, {action: "hello", name: '李四'});
-      response.success();
-    }
-  );
-});
-
-AV.Cloud.define('testRun_options_callback', function(request, response) {
-  AV.Cloud.run('choice', {choice: true}).then(
-    function(data) {
-      assert.equal('OK~', data);
-      AV.Cloud.run('choice', {choice: false}).then(
-        function(data) {
-          assert.ifError(data);
-        },
-        function(err) {
-          assert.equal('OMG...', err);
-          response.success();
-        }
-      );
-    },
-    function(err) {
-      assert.ifError(err);
-    }
-  );
-});
-
-AV.Cloud.define('testRun_promise', function(request, response) {
-  AV.Cloud.run('choice', {choice: true}).then(function(data) {
-    assert.equal('OK~', data);
-    AV.Cloud.run('choice', {choice: false}).then(function(data) {
-      assert.ifError(data);
-    }, function(err) {
-      assert.equal('OMG...', err);
-      response.success();
-    });
-  },
-  function(err) {
-    assert.ifError(err);
-  });
-});
-
-AV.Cloud.define('testRunWithUser', function(request, response) {
-  AV.Cloud.run('testUser', {}, {
-    user: request.user
-  }).then(function(data) {
-    assert.equal('ok', data);
-    response.success();
-  });
-});
-
-AV.Cloud.define('testRunWithAVObject', function(request, response) {
-  AV.Cloud.run('complexObject', {}, {
-    user: request.user
-  }).then(function(datas) {
-    response.success(datas);
-  });
-});
-
-AV.Cloud.define('testRunWithSessionToken', function(request, response) {
-  AV.Cloud.run('testUser', {}, {
-    sessionToken: request.sessionToken
-  }).then(function(datas) {
-    response.success(datas);
-  });
-});
-
-AV.Cloud.define('testRpcRemote', function(request, response) {
-  AV.Cloud.rpc('testRun', {shouldRemote: true}, {
-    remote: true,
-  }).then(function(datas) {
-    response.success(datas);
-  });
-});
-
-AV.Cloud.define('readDir', function(request, response) {
-  fs.readdir('.', function(err, dir) {
-    dir.should.containEql('package.json');
-    response.success(dir);
-  });
-});
-
-AV.Cloud.onVerified('sms', function(request) {
-  assert.equal(request.object.className, '_User');
-  assert.equal(request.object.id, '54fd6a03e4b06c41e00b1f40');
-  assert.equal(request.object.get('username'), 'admin');
-});
-
-AV.Cloud.define('testThrowError', function(request, response) {
-  /* jshint ignore:start */
-  noThisMethod();
-  /* jshint ignore:end */
-  response.success();
-});
-
-AV.Cloud.define("userMatching", function(req, res) {
-  setTimeout(function() {
-    // 为了更加靠谱的验证串号问题，走一次网络 IO
-    var query = new AV.Query(TestObject);
-    query.get('55069f5be4b0c93838ed9b17').then(
-      function(obj) {
-        assert.equal(obj.get('foo'), 'bar');
-        res.success({reqUser: req.user, currentUser: AV.User.current()});
-      }, function() {
-        res.success({reqUser: req.user, currentUser: AV.User.current()});
-      }
-    );
-  }, Math.floor((Math.random() * 2000) + 1));
-});
-
-AV.Cloud.define('testTimeout', function(req, res) {
-  setTimeout(function() {
-    res.success('ok');
-  }, req.params.delay);
-});
-
-AV.Cloud.define('_messageReceived', function(request, response) {
-  response.success('ok');
-});
-
-AV.Insight.on('end', function(err, result) {
-  assert.deepEqual({
-    "id" : "job id",
-    "status": "OK/ERROR",
-    "message": "当 status 为 ERROR 时的错误消息"
-  }, _.omit(result, '__sign'));
-});
-
-var sessionToken_admin = config.sessionToken_admin;
+const app = utils.app();
 
 describe('functions', function() {
   it('ping', function(done) {
@@ -544,7 +294,7 @@ describe('functions', function() {
       .post('/1/functions/testUser')
       .set('X-AVOSCloud-Application-Id', appId)
       .set('X-AVOSCloud-Application-Key', appKey)
-      .set('x-avoscloud-session-token', sessionToken_admin)
+      .set('x-avoscloud-session-token', sessionTokenAdmin)
       .expect(200, done);
   });
 
@@ -553,7 +303,7 @@ describe('functions', function() {
       .post('/1/functions/dontFetchUser')
       .set('X-AVOSCloud-Application-Id', appId)
       .set('X-AVOSCloud-Application-Key', appKey)
-      .set('x-avoscloud-session-token', sessionToken_admin)
+      .set('x-avoscloud-session-token', sessionTokenAdmin)
       .expect(200, done);
   });
 
@@ -577,7 +327,7 @@ describe('functions', function() {
       .post('/1/functions/testRunWithUser')
       .set('X-AVOSCloud-Application-Id', appId)
       .set('X-AVOSCloud-Application-Key', appKey)
-      .set('x-avoscloud-session-token', sessionToken_admin)
+      .set('x-avoscloud-session-token', sessionTokenAdmin)
       .expect(200, done);
   });
 
@@ -587,7 +337,7 @@ describe('functions', function() {
       .post('/1/functions/testRun_options_callback')
       .set('X-AVOSCloud-Application-Id', appId)
       .set('X-AVOSCloud-Application-Key', appKey)
-      .set('x-avoscloud-session-token', sessionToken_admin)
+      .set('x-avoscloud-session-token', sessionTokenAdmin)
       .expect(200, done);
   });
 
@@ -597,7 +347,7 @@ describe('functions', function() {
       .post('/1/functions/testRun_promise')
       .set('X-AVOSCloud-Application-Id', appId)
       .set('X-AVOSCloud-Application-Key', appKey)
-      .set('x-avoscloud-session-token', sessionToken_admin)
+      .set('x-avoscloud-session-token', sessionTokenAdmin)
       .expect(200, done);
   });
 
@@ -606,7 +356,7 @@ describe('functions', function() {
       .post('/1/functions/testRunWithSessionToken')
       .set('X-AVOSCloud-Application-Id', appId)
       .set('X-AVOSCloud-Application-Key', appKey)
-      .set('x-avoscloud-session-token', sessionToken_admin)
+      .set('x-avoscloud-session-token', sessionTokenAdmin)
       .expect(200, done);
   });
 
@@ -665,7 +415,7 @@ describe('functions', function() {
   });
 
   it('timeoutTest', function(done) {
-    this.timeout(17000);
+    this.timeout(20000);
     request(app)
       .post('/1.1/functions/testTimeout')
       .set('X-AVOSCloud-Application-Id', appId)
@@ -713,7 +463,7 @@ describe('functions', function() {
       });
     };
     for (var i = 0; i <= 4; i++) {
-      doRequest(sessionToken_admin, 'admin', cb);
+      doRequest(sessionTokenAdmin, 'admin', cb);
       doRequest('0hgr13u12tmgyv4x594682sv5', 'zhangsan', cb);
       doRequest(null, null, cb);
     }
