+++ test/koa/https-redirect-test.js
@@ -5,9 +5,7 @@ var request = require('supertest');
 require('should');
 
 var AV = require('../..');
-var config = require('../config');
-
-AV.init(config);
+const appInfo = require('../utils/app-info');
 
 var app = koa();
