+++ test/utils/app-info.js
@@ -1,6 +1,12 @@
-module.exports = {
+const AV = require('../..');
+
+const appInfo = {
   appId: '4h2h4okwiyn8b6cle0oig00vitayum8ephrlsvg7xo8o19ne',
   appKey: '3xjj1qw91cr3ygjq9lt0g8c3qpet38rrxtwmmp0yffyoy2t4',
   masterKey: '3v7z633lzfec9qzx8sjql6zimvdpmtwypcchr2gelu5mrzb0',
-  sessionToken_admin: 'w2jrtkbehp38otqmhbqu7ybs9'
+  sessionTokenAdmin: 'w2jrtkbehp38otqmhbqu7ybs9'
 };
+
+AV.init(appInfo);
+
+module.exports = appInfo;
